reorganized files, added router
