+++ schedulesorcerers/src/auth/Login.css
@@ -1,0 +1,129 @@
+.login-page-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    background-image: linear-gradient(#4a5899 10%, #5788bd 70%, #64b6e0);
+  }
+  
+  .left-side-bar {
+    flex: 1.3;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    padding-left: 3%;
+    padding-right: 3%;
+    border-top-left-radius: 20px;
+    border-bottom-left-radius: 20px;
+    background-color: #fff;
+  }
+  
+  .left-side-bar h2 {
+    color: #4a5899;
+    margin-bottom: 20px;
+    text-align: center;
+    font-size: 30px;
+  }
+  
+  .small-line-purple {
+    background-color: #4a5899;
+    width: 70px;
+    height: 2px;
+    margin-top: -10px;
+    margin-bottom: 30px;
+    border-radius: 2px;
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 70%;
+  }
+  
+  .form-group {
+    width: 70%;
+    margin-bottom: 15px;
+  }
+  
+  label {
+    display: block;
+    margin-bottom: 5px;
+  }
+  
+  input {
+    width: calc(100% - 22px);
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 16px;
+  }
+  
+  .left-side-bar button {
+    margin-top: 20px;
+    width: 150px;
+    padding: 10px;
+    border: 1.5px solid #4a5899;
+    border-radius: 20px;
+    background-color: #4a5899;
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+  }
+  
+  button:hover {
+    background-color: #fff;
+    color: #4a5899;
+  }
+  
+  .right-side-bar {
+    background-image: linear-gradient(#4a5899 10%, #5788bd 70%, #64b6e0);
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    padding-left: 3%;
+    padding-right: 3%;
+  }
+  
+  .right-side-bar h2 {
+    color: #fff;
+    text-align: center;
+    font-size: 60px;
+  }
+  
+  .small-line-white {
+    background-color: #fff;
+    width: 70px;
+    height: 2px;
+    margin-top: -30px;
+    border-radius: 2px;
+  }
+  
+  .right-side-bar p1 {
+    color: #fff;
+    text-align: center;
+    margin-top: 30px;
+    margin-bottom: 30px;
+  }
+  
+  .right-side-bar button {
+    width: 150px;
+    padding: 10px;
+    border: 1.5px solid #fff;
+    border-radius: 20px;
+    background-color: transparent;
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+  }
+  
+  .right-side-bar button:hover {
+    background-color: #fff;
+    color: #4a5899;
+  }
