+++ schedulesorcerers/src/home/HomePage.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './HomePage.css';
+
+const HomePage = () => {
+    return (
+        <div>
+          <p>calendar here</p>
+        </div>
+      );
+};
+
+export default HomePage
